// src/context/types/enums.ts
export const enum GameActionKind {
    RESTART_MOVES = 'RESTART_MOVES',
    CHANGE_LEVEL = 'CHANGE_LEVEL',
    SET_CARDS = 'SET_CARDS',
    REVEAL_CARD = 'REVEAL_CARD',
    SET_ACTIVE_CARD = 'SET_ACTIVE_CARD',
    SET_DONE_CARDS = 'SET_DONE_CARDS',
    HIDE_CARDS = 'HIDE_CARDS',
    START_CHECKING_MOVE = 'START_CHECKING_MOVE',
    INCREASE_TIME = 'INCREASE_TIME',
    MARK_MOVE_TO_FAIL = 'MARK_MOVE_TO_FAIL',
    MARK_MOVE_TO_DONE = 'MARK_MOVE_TO_DONE'
}